move futures and neqassign out of yewtil
