+++ examples/futures/src/main.rs
@@ -7,7 +7,7 @@ use wasm_bindgen::JsCast;
 use wasm_bindgen_futures::JsFuture;
 use web_sys::{Request, RequestInit, RequestMode, Response};
 use yew::{html, Component, ComponentLink, Html, ShouldRender};
-use yewtil::future::LinkFuture;
+use yew::utils::future::LinkFuture;
 
 mod markdown;
