+++ packages/yew-agent/Cargo.toml
@@ -15,7 +15,7 @@ serde = { version = "1.0", features = ["derive"] }
 slab = "0.4"
 wasm-bindgen = "0.2"
 yew = { path = "../yew" }
-
+wasm-bindgen-futures = "0.4"
 
 [dependencies.web-sys]
 version = "0.3"
