+++ packages/yew/src/utils/mod.rs
@@ -1,9 +1,13 @@
 //! This module contains useful utilities to get information about the current document.
 
-use anyhow::{anyhow, Error};
 use std::marker::PhantomData;
+
+use anyhow::{anyhow, Error};
 use web_sys::{Document, Window};
+
 use yew::html::ChildrenRenderer;
+
+pub mod not_equal_assign;
 
 /// Returns the current window. This function will panic if there is no available window.
 pub fn window() -> Window {
