+++ packages/yewtil/src/lib.rs
@@ -12,28 +12,14 @@
 //! * "store" - Global state with easy binding
 // //! * "dsl" - Use functions instead of Yew's `html!` macro.
 
-#[cfg(feature = "neq")]
-mod not_equal_assign;
-
-#[cfg(feature = "pure")]
-mod pure;
-
-#[cfg(any(feature = "mrc_irc", feature = "lrc"))]
-pub mod ptr;
-
-#[cfg(feature = "history")]
-mod history;
-
+#[deprecated]
+#[cfg(feature = "effect")]
+pub use effect::{effect, Effect};
 #[cfg(feature = "history")]
 pub use history::History;
-
-#[cfg(feature = "neq")]
-pub use not_equal_assign::{NeqAssign, NeqAssignBy};
-
 #[deprecated]
 #[cfg(feature = "pure")]
 pub use pure::{Pure, PureComponent};
-
 /// Define a component with a function
 ///
 /// # Example
@@ -60,17 +46,20 @@ pub use pure::{Pure, PureComponent};
 #[cfg(feature = "pure")]
 pub use yewtil_macro::function_component;
 
+#[cfg(feature = "pure")]
+mod pure;
+
+#[cfg(any(feature = "mrc_irc", feature = "lrc"))]
+pub mod ptr;
+
+#[cfg(feature = "history")]
+mod history;
+
 #[cfg(feature = "fetch")]
 pub mod fetch;
 
 #[cfg(feature = "effect")]
 mod effect;
-#[deprecated]
-#[cfg(feature = "effect")]
-pub use effect::{effect, Effect};
-
-#[cfg(feature = "future")]
-pub mod future;
 
 #[cfg(feature = "store")]
 pub mod store;
